BSM-122: Include BSM Daemon in default/service-configuration.xml
