+++ core/upgrade/src/test/java/org/opennms/upgrade/implementations/ServiceConfigMigratorOfflineTest.java
@@ -1,8 +1,8 @@
 /*******************************************************************************
  * This file is part of OpenNMS(R).
  *
- * Copyright (C) 2014 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2014 The OpenNMS Group, Inc.
+ * Copyright (C) 2015 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2015 The OpenNMS Group, Inc.
  *
  * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
  *
@@ -28,24 +28,53 @@
 
 package org.opennms.upgrade.implementations;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+
 import java.io.File;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameters;
 import org.opennms.core.utils.ConfigFileConstants;
 import org.opennms.core.xml.JaxbUtils;
 import org.opennms.netmgt.config.ServiceConfigFactory;
+import org.opennms.netmgt.config.service.Service;
 import org.opennms.netmgt.config.service.ServiceConfiguration;
 
 /**
  * The Test Class for ServiceConfigMigratorOffline.
  * 
- * @author Alejandro Galue <agalue@opennms.org>
+ * @author <a href="mailto:ranger@opennms.org">Benjamin Reed</a>
  */
+@RunWith(Parameterized.class)
 public class ServiceConfigMigratorOfflineTest {
+    private final String m_testFile;
+    private final int m_totalBefore;
+    private final int m_totalAfter;
+    private final int m_enabledAfter;
+
+    private final List<String> m_disabled = Arrays.asList(
+            "OpenNMS:Name=Linkd",
+            "OpenNMS:Name=Xmlrpcd",
+            "OpenNMS:Name=XmlrpcProvisioner",
+            "OpenNMS:Name=AccessPointMonitor"
+    );
+
+    public ServiceConfigMigratorOfflineTest(final String testFile, final int totalBefore, final int totalAfter, final int enabledAfter) {
+        m_testFile = testFile;
+        m_totalBefore = totalBefore;
+        m_totalAfter = totalAfter;
+        m_enabledAfter = enabledAfter;
+    }
 
     /**
      * Sets up the test.
@@ -56,6 +85,7 @@ public class ServiceConfigMigratorOfflineTest {
     public void setUp() throws Exception {
         FileUtils.copyDirectory(new File("src/test/resources/etc"), new File("target/home/etc"));
         System.setProperty("opennms.home", "target/home");
+        FileUtils.copyFile(new File(m_testFile), new File("target/home/etc/service-configuration.xml"));
     }
 
     /**
@@ -68,28 +98,52 @@ public class ServiceConfigMigratorOfflineTest {
         FileUtils.deleteDirectory(new File("target/home"));
     }
 
+    @Parameters
+    public static Collection<Object[]> params() {
+        return Arrays.asList(new Object[][] {
+            // service config, total, enabled
+            { "target/home/etc/service-configuration-1.8.17.xml",  38, 29, 29 },
+            { "target/home/etc/service-configuration-1.10.14.xml", 38, 29, 29 },
+            { "target/home/etc/service-configuration-1.12.9.xml",  39, 29, 29 },
+            { "target/home/etc/service-configuration-14.0.3.xml",  38, 30, 24 },
+            { "target/home/etc/service-configuration-15.0.2.xml",  38, 30, 24 },
+            { "target/home/etc/service-configuration-16.0.4.xml",  37, 30, 24 }
+        });
+    }
+
     /**
      * Test fixing the configuration file.
      *
      * @throws Exception the exception
      */
     @Test
-    public void testFixingConfig() throws Exception {
-        ServiceConfigMigratorOffline migrator = new ServiceConfigMigratorOffline();
-        migrator.execute();
-
-        // Checking parsing the fixed file (it should contain all the services)
+    public void testUpgradeConfig() throws Exception {
         File cfgFile = ConfigFileConstants.getFile(ConfigFileConstants.SERVICE_CONF_FILE_NAME);
         ServiceConfiguration cfg = JaxbUtils.unmarshal(ServiceConfiguration.class, cfgFile);
 
-        // Do not change this value: the config file behind this test should always contain
-        // the content for OpenNMS 14.0.0, regardless of whether or not services are added,
-        // changed, or removed in the latest version.
-        Assert.assertEquals(38, cfg.getServiceCount());
+        // check the total and active services before doing the upgrade
+        assertEquals(m_totalBefore, cfg.getServiceCount());
 
-        // Checking Service Factory (it should return only the enabled services)
-        ServiceConfigFactory factory = new ServiceConfigFactory();
-        Assert.assertEquals(27, factory.getServices().length);
+        // perform the upgrade
+        final ServiceConfigMigratorOffline migrator = new ServiceConfigMigratorOffline();
+        migrator.execute();
+
+        // confirm the total and active services after the upgrade
+        cfg = JaxbUtils.unmarshal(ServiceConfiguration.class, cfgFile);
+        final ServiceConfigFactory factory = new ServiceConfigFactory();
+        Assert.assertEquals(m_totalAfter, cfg.getServiceCount());
+        Assert.assertEquals(m_enabledAfter, factory.getServices().length);
+
+        for (final Service svc : cfg.getServiceCollection()) {
+            final String serviceName = svc.getName();
+            if (m_disabled.contains(serviceName)) {
+                assertFalse("Service " + serviceName + " should be disabled.", svc.isEnabled());
+            }
+        }
+        for (final Service svc : factory.getServices()) {
+            final String serviceName = svc.getName();
+            assertFalse("Service " + serviceName + " should not be in the active service list.", m_disabled.contains(svc.getName()));
+        }
     }
 
 }
